# perf(90kernel-modules): use AWK instead of shell monster