@LaszloGombos 2022-10-13
> Not that it's by much, so it's editor's choice, really.
In my system I do not have awk in my generated initramfs (by dracut), so this change breaks my boot.

There are ways to fix this, but adding awk and its dependencies might also make my generated initramfs larger and as such loading it to ram might take extra time (depending on the storage read spead). I do not think you have accounted for that in your micro benchmark.

I do not think we should make this change.