@LaszloGombos 2022-10-13
> I continue to fail to see where you get the idea that this ends up in the initrd.

I just realized as well, so I delated my earlier comment as it is not relevant, sorry about that. I will try to take a closer look later, but this PR seems like a good optimization.