@aafeijoo-suse 2022-10-14
Thanks for finding this. I'd use `grep` to avoid having to add a require to `awk` in the spec file... but it seems upstream is going to drop the spec file sooner or later, and we already implicitly require `awk` since https://github.com/dracutdevs/dracut/commit/9582f027