@LaszloGombos 2022-10-17
> I'd use grep to avoid having to add a require to awk

@nabijaczleweli, would you consider measuring grep as well to make a decision between grep and awk for this optimization ?