@Conan-Kudo 2022-10-17
> Thanks for finding this. I'd use `grep` to avoid having to add a require to `awk` in the spec file... but it seems upstream is going to drop the spec file sooner or later, and we already implicitly require `awk` since [9582f027](https://github.com/dracutdevs/dracut/commit/9582f027)

I'd prefer to keep the spec file, but it needs to be updated and actively tested (which it currently is not).