@LaszloGombos 2022-10-17
> I did; it's benchmark 2 in OP

Sorry. missed it. Thank you !