@pvalena 2022-10-21
> > Thanks for finding this. I'd use `grep` to avoid having to add a require to `awk` in the spec file... but it seems upstream is going to drop the spec file sooner or later, and we already implicitly require `awk` since [9582f027](https://github.com/dracutdevs/dracut/commit/9582f027)
> 

I'm rather in favort of dropping awk as well....

> I'd prefer to keep the spec file, but it needs to be updated and actively tested (which it currently is not).

I'll send PR with backports from Fedora (there's not too much difference). AFAIK we actually use it without any modifications in c9s (and therefore RHEL9):

https://gitlab.com/redhat/centos-stream/rpms/dracut/-/blob/c9s/dracut.spec