@nabijaczleweli 2022-10-22
Wait so let me get this straight: everyone likes this approach except fedora maintainers who would rather spew pages here than fix their packages to include a dependency on literally any AWK implementation, *that dracut already uses in base host-side code*, and their inability to add `Depends: awk` or whatever the fedora spelling is means they vetoed this somehow? Does fedora even *ship* AWK that this is somehow a problem?