@nabijaczleweli 2022-10-22
Yes, you may also notice that that patch is trivially broken.

Here's popcon stats for mawk vs gawk installations in Debian and its derivatives (https://qa.debian.org/popcon-graph.php?packages=dracut+dracut-core+mawk+nawk+gawk&show_installed=on&want_percent=on&want_legend=on&want_ticks=on&from_date=&to_date=&hlght_date=&date_fmt=%25Y-%25m&beenhere=1; untick percent for first):
![image](https://user-images.githubusercontent.com/6709544/197338800-cb6e8930-8bce-40a9-a9f3-fdd92af74f48.png)
![image](https://user-images.githubusercontent.com/6709544/197338839-fb05a45a-8c96-4bc7-bab5-bd5b7ebf77b3.png)

Here we can see that, in a.d. 2022, the decided majority of users has mawk installed but not gawk installed, and this has been the trend since ~2007 (for those that have both, gawk takes precedence as awk by default (Priority: 10 for gawk and Priority: 5 for mawk), so that's the upper limit on gawk use – users are free to pick mawk and I know anecdotally that I and the users I know do; there's no alternatives data reported by popcon).

Given the above, there appears to be 0 reason to fret about Pulling In An AWK, because all systems have one, and very little reason to expect Most Users to have awk=gawk instead of awk=mawk.